Fixed global context axios calls to the server and relocated api.js to server folder
