+++ server/router/reviews.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const api = require('../../src/api.js');
+const api = require('../api.js');
 
 
 router.get('/', (req, res) => {
