+++ src/ProductDetail.jsx
@@ -11,10 +11,11 @@ import { Overview } from './components/Overview/Overview.jsx'
 const ProductDetail = () => {
   let { id } = useParams();
   // console.log('product id from url params', id)
-  const { updateCurrentProductId, currentProductId } = useContext(GlobalContext);
+  const { updateCurrentProductId, currentProductId, getAllProducts } = useContext(GlobalContext);
 
   useEffect(() =>{
     updateCurrentProductId(id);
+    getAllProducts();
     console.log('===> get into the useEffect to update product id')
   },[id])
